% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration.R
\name{configureHMC}
\alias{configureHMC}
\title{Create HMC configuration}
\usage{
configureHMC(model, nodes = character(), control = list(), print = TRUE, ...)
}
\arguments{
\item{model}{A nimble model, as returned by `nimbleModel`}

\item{nodes}{A character vector of stochastic node names to sample by HMC. If an empty character vector is provided (the default), then all non-data nodes will be sampled.}

\item{control}{Optional named list of control parameters to be passed as the `control` argument to `sampler_HMC`.  See `help(sampler_HMC)` for details.}

\item{print}{Logical argument specifying whether to print the montiors and samplers.  Default is TRUE.}

\item{...}{Other arguments that will be passed to `configureMCMC`}
}
\description{
Create a nimble MCMC configuration object containing only an HMC sampler
}
\details{
This function can be used like `configureMCMC` in nimble to create an MCMC configuration object.  It will return an MCMC configuration with only an HMC sampler.

Use this function if you want to create an MCMC configuration and the modify it before building it.  If you want to go straight to an MCMC containing just an HMC sampler, you can us `buildHMC` instead.
}
\examples{
#' nimbleOptions(enableDerivs = TRUE)

code <- nimbleCode({
    b0 ~ dnorm(0, 0.001)
    b1 ~ dnorm(0, 0.001)
    sigma ~ dunif(0, 10000)
    for(i in 1:N) {
        mu[i] <- b0 + b1 * x[i]
        y[i] ~ dnorm(mu[i], sd = sigma)
    }
})

N <- 10
constants <- list(N = N, x = 1:N)
data <- list(y = 1:N)
inits <- list(b0 = 1, b1 = 0.1, sigma = 1)

Rmodel <- nimbleModel(code, constants, data, inits, buildDerivs = TRUE)

## create MCMC configuration object with only an HMC sampler
conf <- configureHMC(Rmodel)

Rmcmc <- buildMCMC(conf)

# Cmodel <- compileNimble(Rmodel)
# Cmcmc <- compileNimble(Rmcmc, project = Rmodel)
# samples <- runMCMC(Cmcmc)
}
\seealso{
\code{\link{addHMC}} \code{\link{buildHMC}} \code{\link{sampler_HMC}} \code{\link{configureMCMC}} \code{\link{addSampler}}
}
\author{
Daniel Turek
}
