% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempSampler.R
\name{myTempSampler}
\alias{myTempSampler}
\title{myTempSampler name}
\usage{
myTempSampler(model, mvSaved, target, control)
}
\arguments{
\item{model}{A nimble model object.  See details.}

\item{mvSaved}{A nimble modelValues object.  See details.}

\item{target}{A character vector of node names.  See details.}

\item{control}{A named list of sampler control elements.  See details.}
}
\description{
myTempSampler description of what it does
}
\details{
XXXX DESCRIBE ARGUMENTS AND SAMPLER USE
}
\examples{

code <- nimbleCode({
    b0 ~ dnorm(0, 0.001)
    b1 ~ dnorm(0, 0.001)
    sigma ~ dunif(0, 10000)
    for(i in 1:N) {
        mu[i] <- b0 + b1 * x[i]
        y[i] ~ dnorm(mu[i], sd = sigma)
    }
})
 
constants <- list(N = 10)
Rmodel <- nimbleModel(code, constants)
conf <- configureMCMC(Rmodel)
conf$setSamplers()
conf$addSampler(target = c('b0', 'b1', 'sigma'), type = 'RW_block')  ## XXXXXX MAKE 'HMC'
Rmcmc <- buildMCMC(conf)

}
\author{
Daniel Turek
}
