% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplers.R
\docType{data}
\name{stateNL_NUTS}
\alias{stateNL_NUTS}
\title{nimbleList definition used internally in NUTS sampler and not expected to be called directly.}
\format{
An object of class \code{list} of length 1.
}
\usage{
stateNL_NUTS
}
\description{
nimbleList definition used internally in NUTS sampler and not expected to be called directly.
}
\keyword{datasets}
