% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMC_samplers.R
\docType{data}
\name{treebranchNL_NUTS}
\alias{treebranchNL_NUTS}
\title{nimbleList definition used internally in NUTS sampler.}
\format{
An object of class \code{list} of length 1.
}
\usage{
treebranchNL_NUTS
}
\description{
nimbleList definition used internally in NUTS sampler.
}
\keyword{datasets}
