% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configureHMC.R
\name{configureHMC}
\alias{configureHMC}
\title{Create HMC configuration}
\usage{
configureHMC(model, nodes = character(), control = list(), ...)
}
\arguments{
\item{model}{A nimble model, such as returned by `nimbleModel`.}

\item{nodes}{A character vector of stochastic node names to sample by HMC.
If nodes is `character()`, all nodes will be sampled.}

\item{control}{Optional list of control parameters to be passed as the `control`.
argument to `sampler_HMC`.  See `help(sampler_HMC)`.}

\item{...}{Other arguments that will be passed to `configureMCMC`.}
}
\description{
Create a nimble MCMC configuration object containing only an HMC
}
\details{
This function can be used like `configureMCMC` in nimble.  It will return
an MCMC configuration with only an HMC sampler.

Use this function if you want to create an MCMC configuration and the modify it
before building it.  If you want to go straight to a built MCMC sampler containing
just an HMC sampler, use buildMCMC instead.
}
\examples{
# See example for help(addHMC)
# Replace the configureMCMC and addHMC lines with
# conf <- configureHMC(Rmodel)
}
